+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Ability to list your hook secrets via `hook secrets`.
 - Add the ID and WorkspaceID to the secret output.
+- Ability to create a workspace via `workspace create`.
 
 ## 0.0.12